cda/requirements.txt
minimalmodbus
paho-mqtt
pyserial
RPi.GPIO